requirements.txt
# For handling data
pandas==2.0.3

# For vectorization and machine learning models
scikit-learn==1.3.0

# For text processing
nltk==3.8.1
regex==2024.1.1

# For reading PDFs
pypdf==3.7.0

# For reading DOCX files
python-docx==0.8.11

# For web application development
streamlit==1.25.0

# For visualization
matplotlib==3.7.1
seaborn==0.12.2

# For saving and loading models
pickle5==0.0.12
